requirements.txt
attrs==25.3.0
Automat==24.8.1
botocore>=1.29.0
boto3>=1.20.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
cryptography==44.0.2
cssselect==1.3.0
defusedxml==0.7.1
docker==7.1.0
fake-useragent==2.1.0
filelock==3.18.0
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==5.3.1
outcome==1.3.0.post0
packaging==24.2
parsel==1.10.0
Protego==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==25.0.0
PySocks==1.7.1
python-dotenv==1.1.0
PyYAML==6.0.2
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
retrying==1.3.4
scrapinghub==2.5.0
scrapy==2.6.3
scrapy-selenium==0.0.7
selenium==4.9.1
service-identity==24.2.0
setuptools==77.0.3
shub==2.15.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tldextract==5.1.3
toml==0.10.2
tqdm==4.55.1
trio==0.29.0
trio-websocket==0.12.2
Twisted==24.11.0
typing_extensions==4.13.0
urllib3<2.1,>=1.25.4
websocket-client==1.8.0
wsproto==1.2.0
zope.interface==7.2
